сменил int на long для параметра конфига BoardId

+++ DirectumRXAutoDeployer/Notifiers/AgileBoards/AgileBoardSettings.cs
@@ -5,7 +5,7 @@ namespace DirectumRXAutoDeployer.Notifiers.AgileBoards
     public class AgileBoardSettings
     {
         public string Token { get; set; }
-        public int BoardId { get; set; }
+        public long BoardId { get; set; }
         public string IntegrationServiceUri { get; set; }
         public string ColumnFrom { get; set; }
         public string ColumnTo { get; set; }
